drf_clientes/requirements.txt
asgiref==3.8.1
astroid==3.2.2
colorama==0.4.6
dill==0.3.8
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
Faker==25.9.1
isort==5.13.2
lazy-object-proxy==1.10.0
Markdown==3.6
mccabe==0.7.0
platformdirs==4.2.2
pylint==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
tomlkit==0.12.5
tzdata==2024.1
validate-docbr==1.10.0
wrapt==1.16.0
